General refinement
Insertion of official content in the database script (data was gathered
by Andreas Rigopoulos), Code improvements, Addition of appropriate
comments.

+++ getLost - Beta/admin/Delete_Data_Locations.php
@@ -1,13 +1,17 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
-<html>
-<body>
 <?php
-	require_once 'Connect_DB.php';
-    require_once 'dbHandler.php';
-	$db = new dbHandler();
-	$id=$_GET['id'];
-	$db->delete("LOCATION", array("LID" => $id));
-	header('location:View_Locations.php');
+// PHP code to execute a delete query on the LOCATION table
+	if (isset($_GET['id']))
+	{
+		require_once 'Connect_DB.php';
+		require_once 'dbHandler.php';
+		$db = new dbHandler();             // Creation of the object  
+		
+		$id = strip_tags($_GET['id']);
+		$db->delete("LOCATION", array("LID" => $id));   // Delete the specified record
+		header('location:View_Locations.php');    
+	}
+	else
+	{
+		echo "An error occurred, go back and try again!";
+	}
 ?>
-</body>
-</html>
