+++ getLost - Beta/admin/Delete_Data_Places.php
@@ -1,13 +1,17 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01//EN" "http://www.w3.org/TR/html4/strict.dtd">
-<html>
-<body>
 <?php
-	require_once 'Connect_DB.php';
-    require_once 'dbHandler.php';
-	$db = new dbHandler();
-	$id=$_GET['id'];
-	$db->delete("PLACE", array("PID" => $id));
-	header('location:View_Places.php');
+// PHP code to execute a delete query on the PLACE table
+	if (isset($_GET['id']))
+	{
+		require_once 'Connect_DB.php';
+		require_once 'dbHandler.php';
+		$db = new dbHandler();         // Creation of the object
+		
+		$id=strip_tags($_GET['id']); 
+		$db->delete("PLACE", array("PID" => $id));    // Delete the specified record
+		header('location:View_Places.php');   
+	}
+	else
+	{
+		echo "An error occurred, go back and try again!";
+	}
 ?>
-</body>
-</html>
